+++ NEWS.md
@@ -5,6 +5,7 @@
 - Thanks to @tjmckinley, a bug in `update()` when the `object` is an instance of the `dgp` class (that has been trimmed by `window()`) is fixed.
 - Thanks to @tjmckinley, some R memory issues due to the underlying Python implementations are rectified.
 - `set_seed()` function is added to ensure reproducible results from the package.
+- A bug is fixed when candidate sets `x_cand` and `y_cand` are provided to `design()`.
 
 # dgpsi 2.1.6